提取信号成功，可以绘制频谱数据， 存在问题： 退出程序会死掉（可能是线程没有退出）

+++ app/src/main/java/com/ldkj/illegal_radio/utils/devices/base/IDevice.java
@@ -11,7 +11,7 @@ public interface IDevice {
     boolean sendCMD(String pCMD) ;
 
     double getLevel();
-    short[] getScanDate();
-    short[] getSpecDate();
+    float[] getScanDate();
+    float[] getSpecDate();
 
 }
